MNER/requirements.txt
faiss==1.5.3
filelock==3.6.0
numpy==1.21.5
Pillow==9.4.0
psutil==5.8.0
py3nvml==0.2.7
pytorch_crf==0.7.2
pytorch_lightning==1.3.1
requests==2.28.1
scikit_learn==1.2.0
seqeval==1.2.2
tensorflow==2.11.0
torch==1.11.0+cu113
torch_xla==1.0
torchcrf==1.1.0
torchvision==0.12.0+cu113
tqdm==4.64.1
transformers==4.11.3